% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vephelpers.R
\name{cdna_to_transcript_length}
\alias{cdna_to_transcript_length}
\title{Convert cDNA_position to numeric indication of transcript length}
\usage{
cdna_to_transcript_length(cdna)
}
\arguments{
\item{cdna}{vector representing cDNA_position VEP annotations.}
}
\value{
numeric transcript length passed from VEP cDNA_position
}
\description{
This function just pulls out digits from the cDNA_position column.
Note transcript length isn't encoded into cDNA position column UNLESS --total_length flag is on when running vep (turns format into cdna_position/total_length)
It is not possible to turn this flag on using VEP online. If we can't find total length - we just return 0, so for online version we just won't have a transcript sort.
}
