% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vephelpers.R
\name{choose_a_single_transcript_and_consequence}
\alias{choose_a_single_transcript_and_consequence}
\title{Take}
\usage{
choose_a_single_transcript_and_consequence(biotypes, consequence, cdna)
}
\arguments{
\item{biotypes}{transcript biotype (e.g. ) (character)}

\item{cdna}{}

\item{transcript_id}{(character)}
}
\description{
Take
}
