% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2maf.R
\name{df2maf}
\alias{df2maf}
\title{Convert VCF-like data.frame to MAF}
\usage{
df2maf(
  data,
  ref_genome,
  keep_all = TRUE,
  col_chrom = "chr",
  col_pos = "pos",
  col_sample_identifier = "sample",
  col_ref = "ref",
  col_alt = "alt",
  col_consequence = "consequence",
  consequence_dictionary = c("SO", "PAVE", "AUTO"),
  col_gene = "gene",
  col_center = NULL,
  col_entrez_gene_id = NULL,
  col_dbSNP_rsid = NULL,
  col_dbSNP_validation_status = NULL,
  col_matched_normal_sample_identifier = NULL,
  col_sequencer = NULL,
  col_sequence_source = NULL,
  col_mutation_status = NULL
)
}
\arguments{
\item{data}{a data.frame with 1 row per mutation in a cohort (data.frame)}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{keep_all}{keep all columns in the original data.frame? (flag). If FALSE, only includes the minimal required columns and any column explicitly mapped using \verb{col_<name>} arguments.}

\item{col_chrom}{name of column describing chromosome of the mutation (string)}

\item{col_pos}{name of column describing the 1-based position of the mutation (string)}

\item{col_sample_identifier}{name of column describing the sample containing the mutation (string)}

\item{col_ref}{name of column describing the reference allele (string)}

\item{col_alt}{name of column describing the alternate allele (string)}

\item{col_consequence}{name of column describing the consequence of the mutation (in SO ontology terms e.g. those that VEP would use OR PAVE ontology terms) containing the mutation (string)}

\item{consequence_dictionary}{What dictionary is to describe variant consequences (SO / PAVE / etc)}

\item{col_gene}{name of column containing Hugo_Symbol of the gene affected by the mutation (string)}

\item{col_center}{name of column containing the genome sequencing center reporting the variant (string)}

\item{col_entrez_gene_id}{name of column containing entrez gene IDs (string)}

\item{col_dbSNP_rsid}{name of column describing the dbSNP rsid of the variant, or "novel" if there is no dbSNP record (string)}

\item{col_dbSNP_validation_status}{name of column describing the validation status of the variant. Elements must be one of by1000genomes;by2Hit2Allele; byCluster; byFrequency; byHapMap; byOtherPop; bySubmitter; alternate_allele (string)}

\item{col_matched_normal_sample_identifier}{name of column describing the matched normal sample identifier}

\item{col_sequencer}{name of column describing the instrument used to produce data (string)}

\item{col_sequence_source}{name of column describing the molecular assay type used to produce the analytes used for sequencing (string). Elements are usually one of 'WGS', 'WGA', 'WXS', 'RNA-seq', etc}

\item{col_mutation_status}{name of column describing the mutation status (string). Elements must be one of: None, Germline, Somatic, LOH, Post-transcriptional modification, or Unknown}
}
\value{
a maf-like data.frame (data.table)
}
\description{
Converts a VEP-like (chr-pos-ref-alt-consequence-sample style) dataframe with to a minimal MAF dataframe.
\strong{Input dataframe should include the columns:}
\itemize{
\item \strong{chr}
\item \strong{pos} (1-based)
\item \strong{sample}
\item \strong{ref}
\item \strong{alt}
\item \strong{consequence} (Sequence Ontology terms)
\item \strong{gene}
}
}
\details{
Alternate column names can be used, but require \strong{col_chrom}, \strong{col_pos}, \strong{col_\if{html}{\out{<etc>}}} arguments to be supplied.
}
\examples{

# Start with a vcf/maf-like 'chr-pos-ref-alt-consequence-sample' data.frame
df = read.csv(system.file(package = "vcf2mafR", "testfiles/test_so.tsv"), sep = "\t")

# Convert to a MAF dataframe
df2maf(df, ref_genome = "hg38")

}
