% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{paths_to_sample_names}
\alias{paths_to_sample_names}
\title{Extract Sample Names From VCF filepath}
\usage{
paths_to_sample_names(
  filenames,
  extract = c("before_dot", "before_underscore")
)
}
\arguments{
\item{filenames}{paths to vcf files}

\item{extract}{what text do we extract as sample name. See details}

\item{sep_is_regex}{}
}
\value{
sample names extracted from filename (character)
}
\description{
Extract Sample Names From VCF filepath
}
\details{
\tabular{ll}{
   Extraction \tab Explanation \cr
   before_dot \tab \strong{sample_name}.otherinfo.vcf \cr
   before_underscore \tab \strong{samplename}_other_info.vcf \cr
}
}
\examples{
path <- "/path/to/sample_name.otherinfo.vcf"
paths_to_sample_names(path)

}
