% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2maf.R
\name{valid_maf_columns}
\alias{valid_maf_columns}
\title{Maf Column Names}
\usage{
valid_maf_columns()
}
\value{
Names of each  GDC MAF column in the order they appear (character)
}
\description{
List all valid MAF columns based on GDC specification
}
\examples{
valid_maf_columns()

}
