% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2maf}
\alias{vcf2maf}
\title{Convert Annotated VCF to MAF-compatible data.frame}
\usage{
vcf2maf(
  vcf,
  ref_genome,
  tumor_id = vcf_tumor_id,
  normal_id = vcf_normal_id,
  vcf_tumor_id = "TUMOR",
  vcf_normal_id = "NORMAL",
  anno = c("PAVE", "VEP"),
  missing_to_silent = TRUE,
  verbose = TRUE,
  debug_mode = FALSE
)
}
\arguments{
\item{vcf}{path to a VEP-annotated VCF file}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{tumor_id}{desired value of Tumor_Sample_Barcode in maf. By default will use the name of the tumor sample in the VCF (string)}

\item{normal_id}{desired value of Matched_Norm_Sample_Barcode in maf.  By default will use the name of the normal sample in the VCF (string)}

\item{vcf_tumor_id}{the sample ID describing the tumor in the (string)}

\item{vcf_normal_id}{the sample ID describing the normal in the (string)}

\item{anno}{What annotation tool was used. One of PAVE (oncoanalyser default) or VEP?}

\item{missing_to_silent}{Assyne any missing (NA) or empty ('') consequences are 'Silent' mutations}

\item{verbose}{(flag)}

\item{debug_mode}{run in debug mode (flag)}
}
\value{
a maf compatible data.frame
}
\description{
Expects a vepped VCF. VEP must be run with a couple of options:
\enumerate{
\item Identifiers: Gene symbol & Transcript Version & (HGVS: optional)
\item Transcript annotation: Transcript biotype & 'Identify canonical transcripts'
\item (Optional) If using the commandline version of vep also make sure to use --total_length (helps in transcript choice)
}
}
\examples{
path_vcf_vepped <- system.file("testfiles/test_b38.vepgui.vcf", package = "vcf2mafR")
vcf2maf(vcf = path_vcf_vepped, ref_genome = "b38")

}
