% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcfs2maf}
\alias{vcfs2maf}
\title{Convert several VCF files into a cohort-level MAF}
\usage{
vcfs2maf(
  vcfs,
  ref_genome,
  vcf_tumor_id = "TUMOR",
  vcf_normal_id = "NORMAL",
  parse_tumor_id_from_filename = TRUE,
  extract = c("before_dot", "before_underscore"),
  verbose = TRUE
)
}
\arguments{
\item{vcfs}{path to vcf files (character)}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{vcf_tumor_id}{a vector describing what the tumor_names to expect in a VCF are}

\item{vcf_normal_id}{the sample ID describing the normal in the (string)}

\item{parse_tumor_id_from_filename}{should tumour id be parsed from filename (flag)}

\item{extract}{what text do we extract as sample name. See details}

\item{verbose}{(flag)}
}
\value{
a maf-compatible data.frame
}
\description{
Take multiple VCF files, each representing a single tumour-normal file and combine into one big maf
}
\examples{
vcf_filepaths = dir(system.file(package='vcf2mafR', 'testfiles/cohort_of_vcfs/'), full.names = TRUE)
vcfs2maf(vcf_filepaths)
}
