% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vephelpers.R
\name{vep_rank_biotypes}
\alias{vep_rank_biotypes}
\title{Rank VEP biotypes}
\usage{
vep_rank_biotypes(biotypes)
}
\arguments{
\item{biotypes}{A character vector of VEP annotated biotypes.}
}
\value{
A numeric vector where values represent the priority of each biotype (lower number = higher priority).
}
\description{
This function takes a vector of VEP annotated biotypes and returns their ranks based on priority.
}
\examples{
vep_rank_biotypes(c("protein_coding", "miRNA", "pseudogene"))

}
