% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vephelpers.R
\name{vep_rank_consequences}
\alias{vep_rank_consequences}
\title{Rank VEP consequences}
\usage{
vep_rank_consequences(consequences)
}
\arguments{
\item{consequences}{A character vector of VEP annotated consequences.}
}
\value{
A numeric vector where values represent the priority of each consequence (lower number = higher priority).
}
\description{
This function takes a vector of VEP annotated biotypes and returns their ranks based on priority.
If there are multiple consequences in one string (e.g. splice_region_variant&TFBS_ablation) this function will automatically return the priority for the most severe event
}
\examples{
vep_rank_consequences(c("protein_coding", "miRNA", "pseudogene"))

}
