% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{Convert a vepped VCF into a table ready to be passed into df2maf}
\usage{
vcf2df(
  vcf,
  tumor_id = "Sample1Tumor",
  normal_id = "Sample1Normal",
  debug_mode = TRUE
)
}
\arguments{
\item{vcf}{path to a VCF file}

\item{tumor_id}{desired value of Tumor_Sample_Barcode in maf (string)}

\item{normal_id}{desired value of Matched_Norm_Sample_Barcode in maf (string)}
}
\value{
data.frame
}
\description{
Convert a vepped VCF into a table ready to be passed into df2maf
}
\examples{
path_vcf <- system.file(package = "vcf2mafR", "testfiles/test_grch38.vep.vcf")
vcf2df(path_vcf)
}
