% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2maf}
\alias{vcf2maf}
\title{VCF to MAF}
\usage{
vcf2maf(
  vcf,
  ref_genome,
  tumor_id = "TUMOR",
  normal_id = "NORMAL",
  vcf_tumor_id = tumor_id,
  vcf_normal_id = normal_id,
  verbose = TRUE,
  debug_mode = FALSE
)
}
\arguments{
\item{vcf}{path to a VEP-annotated VCF file}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{tumor_id}{desired value of Tumor_Sample_Barcode in maf (string)}

\item{normal_id}{desired value of Matched_Norm_Sample_Barcode in maf (string)}

\item{vcf_tumor_id}{the sample ID describing the tumor in the (string)}

\item{vcf_normal_id}{the sample ID describing the normal in the (string)}

\item{verbose}{(flag)}
}
\value{
a maf compatible data.frame
}
\description{
Convert a vep-annotated VCF file to a MAF-compatible data.frame
}
\examples{
path_vcf_vepped <- system.file("testfiles/test_b38.vepgui.vcf", package = "vcf2mafR")
vcf2df(vcf = path_vcf_vepped)

}
